backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "uvicorn",
    "authlib",
    "auth0-fastapi-api",
    "langchain-google-genai",
    "tortoise-orm",
    "aiosqlite",
    "aiosmtplib",
    "gunicorn",
    "aerich",
    "asyncpg",
    "python-dotenv",
    "pydantic-settings",
    "fastadmin",
    "bcrypt",
    "intuit-oauth"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]

[tool.aerich]
tortoise_orm = "db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."